feat(backend): Add https domain during localhost testing feat(nextjs): Add https domain during localhost testing fix(backend): trpc api not compiling due to circular dependencies
