+++ .env.example
@@ -12,4 +12,10 @@ NEXT_PUBLIC_AXIOM_DATASET="DATASET_NAME"
 NEXT_PUBLIC_AXIOM_TOKEN=""
 NEXT_PUBLIC_AXIOM_LOG_LEVEL=off
 
-NEXT_PUBLIC_BACKEND_URL="http://localhost:8787/trpc"
+# Used to have your own domaini for the backend server
+SUBDOMAIN_NAME="skylar-backend-YOUR_NAME"
+NEXT_PUBLIC_BACKEND_URL="https://skylar-backend-YOUR_NAME.loca.it/trpc"
+
+# GOOGLE PROVIDER OAUTH 
+GOOGLE_PROVIDER_CLIENT_ID=""
+GOOGLE_PROVIDER_CLIENT_SECRET=""
