+++ apps/backend/.dev.vars.example
@@ -1,4 +1,4 @@
-APP_URL=http://localhost:3000
+APP_URL=https://skylar-frontend-DEV_NAME.loca.it
 SUPABASE_JWT_SECRET=""
 DATABASE_URL=""
 AXIOM_DATASET=""
