+++ packages/api/index.ts
@@ -3,7 +3,7 @@ import type { inferRouterInputs, inferRouterOutputs } from "@trpc/server";
 import type { AppRouter } from "./src/root";
 
 export { appRouter, type AppRouter } from "./src/root";
-export { createTRPCContext } from "./src/trpc";
+export { createTRPCContext } from "./src/trpc/context";
 
 /**
  * Inference helpers for input types
