--- packages/api/src/router/blank.ts
@@ -1,7 +1,0 @@
-import { createTRPCRouter, publicProcedure } from "../trpc";
-
-export const blankRouter = createTRPCRouter({
-  hello: publicProcedure.query(() => {
-    return "you can see this secret message!";
-  }),
-});
