--- packages/api/src/trpc.ts
@@ -1,138 +1,0 @@
-/**
- * YOU PROBABLY DON'T NEED TO EDIT THIS FILE, UNLESS:
- * 1. You want to modify request context (see Part 1)
- * 2. You want to create a new middleware or type of procedure (see Part 3)
- *
- * tl;dr - this is where all the tRPC server stuff is created and plugged in.
- * The pieces you will need to use are documented accordingly near the end
- */
-import { initTRPC } from "@trpc/server";
-import superjson from "superjson";
-
-import type { Session } from "@skylar/auth";
-import { getSession } from "@skylar/auth";
-import type { DbType } from "@skylar/db";
-import type { Logger } from "@skylar/logger";
-import { formatValidatorError } from "@skylar/schema";
-
-import { enforceUserIsAuthed } from "./middleware/enforce-user-is-authed";
-
-/**
- * 1. CONTEXT
- *
- * This section defines the "contexts" that are available in the backend API
- *
- * These allow you to access things like the database, the session, etc, when
- * processing a request
- *
- */
-type CreateContextOptions = {
-  session?: Session;
-  env: { JWT_SECRET: string };
-  logger: Logger;
-  db: DbType;
-};
-
-/**
- * This helper generates the "internals" for a tRPC context. If you need to use
- * it, you can export it from here
- *
- * Examples of things you may need it for:
- * - testing, so we dont have to mock Next.js' req/res
- * - trpc's `createSSGHelpers` where we don't have req/res
- * @see https://create.t3.gg/en/usage/trpc#-servertrpccontextts
- */
-const createInnerTRPCContext = (opts: CreateContextOptions) => {
-  return {
-    session: opts.session,
-    env: opts.env,
-    db: opts.db,
-    logger: opts.logger,
-  };
-};
-
-/**
- * This is the actual context you'll use in your router. It will be used to
- * process every request that goes through your tRPC endpoint
- * @link https://trpc.io/docs/context
- */
-export const createTRPCContext = async ({
-  req,
-  env,
-  logger,
-  db,
-}: {
-  req: Request;
-  logger: Logger;
-  env: CreateContextOptions["env"];
-  db: DbType;
-}) => {
-  const authHeader = req?.headers.get("Authorization") ?? undefined;
-  const session = await getSession({
-    JWT_SECRET: env.JWT_SECRET,
-    authHeader,
-  });
-  const source = req?.headers.get("x-trpc-source") ?? "unknown";
-
-  logger.info(`>>> tRPC Request from ${source} by ${session?.user?.name}`);
-
-  return createInnerTRPCContext({
-    logger,
-    session,
-    env,
-    db,
-  });
-};
-
-/**
- * 2. INITIALIZATION
- *
- * This is where the trpc api is initialized, connecting the context and
- * transformer
- */
-const t = initTRPC.context<typeof createTRPCContext>().create({
-  transformer: superjson,
-  errorFormatter({ shape, error }) {
-    return {
-      ...shape,
-      data: {
-        ...shape.data,
-        parserError: formatValidatorError(error.cause),
-      },
-    };
-  },
-});
-
-/**
- * 3. ROUTER & PROCEDURE (THE IMPORTANT BIT)
- *
- * These are the pieces you use to build your tRPC API. You should import these
- * a lot in the /src/server/api/routers folder
- */
-
-/**
- * This is how you create new routers and subrouters in your tRPC API
- * @see https://trpc.io/docs/router
- */
-export const createTRPCRouter = t.router;
-export const createMiddleware = t.middleware;
-
-/**
- * Public (unauthed) procedure
- *
- * This is the base piece you use to build new queries and mutations on your
- * tRPC API. It does not guarantee that a user querying is authorized, but you
- * can still access user session data if they are logged in
- */
-export const publicProcedure = t.procedure;
-
-/**
- * Protected (authed) procedure
- *
- * If you want a query or mutation to ONLY be accessible to logged in users, use
- * this. It verifies the session is valid and guarantees ctx.session.user is not
- * null
- *
- * @see https://trpc.io/docs/procedures
- */
-export const protectedProcedure = t.procedure.use(enforceUserIsAuthed);
