+++ packages/api/src/trpc/factory.ts
@@ -1,0 +1,40 @@
+import { initTRPC } from "@trpc/server";
+import superjson from "superjson";
+
+import { formatValidatorError } from "@skylar/schema";
+
+import type { createTRPCContext } from "../..";
+
+/**
+ * 2. INITIALIZATION
+ *
+ * This is where the trpc api is initialized, connecting the context and
+ * transformer
+ */
+const t = initTRPC.context<typeof createTRPCContext>().create({
+  transformer: superjson,
+  errorFormatter({ shape, error }) {
+    return {
+      ...shape,
+      data: {
+        ...shape.data,
+        parserError: formatValidatorError(error.cause),
+      },
+    };
+  },
+});
+
+/**
+ * 3. ROUTER & PROCEDURE (THE IMPORTANT BIT)
+ *
+ * These are the pieces you use to build your tRPC API. You should import these
+ * a lot in the /src/server/api/routers folder
+ */
+
+/**
+ * This is how you create new routers and subrouters in your tRPC API
+ * @see https://trpc.io/docs/router
+ */
+export const createTRPCRouter = t.router;
+export const createMiddleware = t.middleware;
+export const procedure = t.procedure;
